+++ src/metaSMT/backend/Boolector.hpp
@@ -289,14 +289,14 @@ namespace metaSMT {
         }
 
         result_type operator() (bvtags::zero_extend_tag const &
-            , unsigned width
+            , uint64_t width
             , result_type e)
         {
           return ptr(boolector_uext(_btor, e, width));
         }
 
         result_type operator() (bvtags::sign_extend_tag const &
-            , unsigned width
+            , uint64_t width
             , result_type e)
         {
           return ptr(boolector_sext(_btor, e, width));
